SASS 5478 12 month period page & Received Grant Amount Page

+++ app/pages/premiumlease/RecievedGrantLeaseAmountPage.scala
@@ -20,7 +20,7 @@ import pages.QuestionPage
 import play.api.libs.json.JsPath
 
 
-case object RecievedGrantLeaseAmountPage extends QuestionPage[Int] {
+case object RecievedGrantLeaseAmountPage extends QuestionPage[BigDecimal] {
 
   override def path: JsPath = JsPath \ toString
